+++ rust/operator-binary/src/druid_controller.rs
@@ -5,18 +5,20 @@ use crate::{
 };
 
 use snafu::{OptionExt, ResultExt, Snafu};
+use stackable_druid_crd::tls::DruidTlsSettings;
 use stackable_druid_crd::{
-    resource::RoleResource, DeepStorageSpec, DruidCluster, DruidRole, APP_NAME,
-    AUTH_AUTHORIZER_OPA_URI, CERTS_DIR, CONTAINER_HTTP_PORT, CONTAINER_METRICS_PORT,
-    CONTROLLER_NAME, CREDENTIALS_SECRET_PROPERTY, DRUID_CONFIG_DIRECTORY, DRUID_METRICS_PORT,
-    DS_BUCKET, HDFS_CONFIG_DIRECTORY, JVM_CONFIG, LOG4J2_CONFIG, RUNTIME_PROPS,
-    RW_CONFIG_DIRECTORY, S3_ENDPOINT_URL, S3_PATH_STYLE_ACCESS, S3_SECRET_DIR_NAME,
+    authentication, authentication::DruidAuthentication, resource::RoleResource, tls,
+    DeepStorageSpec, DruidAuthorization, DruidCluster, DruidRole, APP_NAME,
+    AUTH_AUTHORIZER_OPA_URI, CERTS_DIR, CONTROLLER_NAME, CREDENTIALS_SECRET_PROPERTY,
+    DRUID_CONFIG_DIRECTORY, DS_BUCKET, HDFS_CONFIG_DIRECTORY, JVM_CONFIG, LOG4J2_CONFIG,
+    RUNTIME_PROPS, RW_CONFIG_DIRECTORY, S3_ENDPOINT_URL, S3_PATH_STYLE_ACCESS, S3_SECRET_DIR_NAME,
     ZOOKEEPER_CONNECTION_STRING,
 };
 use stackable_operator::{
     builder::{
         ConfigMapBuilder, ContainerBuilder, ObjectMetaBuilder, PodBuilder,
-        PodSecurityContextBuilder, SecretOperatorVolumeSourceBuilder, VolumeBuilder,
+        PodSecurityContextBuilder, SecretOperatorVolumeSourceBuilder, SecurityContextBuilder,
+        VolumeBuilder,
     },
     cluster_resources::ClusterResources,
     commons::{
@@ -27,11 +29,9 @@ use stackable_operator::{
     k8s_openapi::{
         api::{
             apps::v1::{StatefulSet, StatefulSetSpec},
-            core::v1::{
-                ConfigMap, EnvVar, Probe, Service, ServicePort, ServiceSpec, TCPSocketAction,
-            },
+            core::v1::{ConfigMap, EnvVar, Service, ServiceSpec},
         },
-        apimachinery::pkg::{apis::meta::v1::LabelSelector, util::intstr::IntOrString},
+        apimachinery::pkg::apis::meta::v1::LabelSelector,
     },
     kube::{
         runtime::{controller::Action, reflector::ObjectRef},
@@ -44,9 +44,9 @@ use stackable_operator::{
     product_config_utils::{transform_all_roles_to_config, validate_all_roles_and_groups_config},
     role_utils::RoleGroupRef,
 };
-use std::ops::Deref;
 use std::{
     collections::{BTreeMap, HashMap},
+    ops::Deref,
     str::FromStr,
     sync::Arc,
     time::Duration,
@@ -183,6 +183,10 @@ pub enum Error {
     NoSegmentCacheLocation,
     #[snafu(display("role group and resource type mismatch. this is a programming error."))]
     RoleResourceMismatch,
+    #[snafu(display("invalid authentication configuration"))]
+    InvalidAuthenticationConfig { source: authentication::Error },
+    #[snafu(display("invalid tls configuration"))]
+    InvalidTlsConfig { source: tls::Error },
 }
 
 type Result<T, E = Error> = std::result::Result<T, E>;
@@ -197,7 +201,7 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
     tracing::info!("Starting reconcile");
     let client = &ctx.client;
 
-    let zk_confmap = druid.spec.zookeeper_config_map_name.clone();
+    let zk_confmap = druid.spec.cluster_config.zookeeper_config_map_name.clone();
     let zk_connstr = client
         .get::<ConfigMap>(&zk_confmap, druid.namespace().as_deref())
         .await
@@ -211,9 +215,11 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
         })?;
 
     // Assemble the OPA connection string from the discovery and the given path, if a spec is given.
-    let opa_connstr = if let Some(opa_spec) = &druid.spec.opa {
+    let opa_connstr = if let Some(DruidAuthorization { opa: opa_config }) =
+        &druid.spec.cluster_config.authorization
+    {
         Some(
-            opa_spec
+            opa_config
                 .full_document_url_from_config_map(
                     client,
                     druid.deref(),
@@ -222,7 +228,7 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
                 )
                 .await
                 .context(GetOpaConnStringSnafu {
-                    cm_name: opa_spec.config_map_name.clone(),
+                    cm_name: opa_config.config_map_name.clone(),
                 })?,
         )
     } else {
@@ -235,7 +241,7 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
         .await
         .context(GetS3ConnectionSnafu)?;
 
-    let deep_storage_bucket_name = match &druid.spec.deep_storage {
+    let deep_storage_bucket_name = match &druid.spec.cluster_config.deep_storage {
         DeepStorageSpec::S3(s3_spec) => {
             s3_spec
                 .bucket
@@ -247,9 +253,34 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
         _ => None,
     };
 
+    // Get possible authentication methods
+    let resolved_authentication_config = DruidAuthentication::resolve(client, &druid)
+        .await
+        .context(InvalidAuthenticationConfigSnafu)?;
+
+    let mut tls_authentication_config = None;
+    let mut authentication_config = vec![];
+    // Remove a TLS provider if available. The TLS provider must be treated in combination
+    // with TLS encryption. Retain all other authentication providers.
+    for auth_conf in resolved_authentication_config.into_iter() {
+        if auth_conf.is_tls_auth() {
+            tls_authentication_config = Some(auth_conf);
+        } else {
+            authentication_config.push(auth_conf);
+        }
+    }
+
+    // Extract TLS encryption and TLS auth provider if available
+    let druid_tls_settings = DruidTlsSettings {
+        encryption: druid.spec.cluster_config.tls.clone(),
+        // There can currently only be one TLS authentication config, so if we find it we can just
+        // take ownership.
+        authentication: tls_authentication_config,
+    };
+
     let role_config = transform_all_roles_to_config(&*druid, druid.build_role_properties());
     let validated_role_config = validate_all_roles_and_groups_config(
-        druid_version(&druid)?,
+        druid.version(),
         &role_config.context(ProductConfigTransformSnafu)?,
         &ctx.product_config,
         false,
@@ -266,7 +297,7 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
             role: role_name.to_string(),
         })?;
 
-        let role_service = build_role_service(role_name, &druid)?;
+        let role_service = build_role_service(&druid, &druid_role, &druid_tls_settings)?;
         cluster_resources
             .add(client, &role_service)
             .await
@@ -282,23 +313,25 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
                 .resources(&druid_role, &rolegroup)
                 .context(FailedToResolveResourceConfigSnafu)?;
 
-            let rg_service = build_rolegroup_services(&rolegroup, &druid, rolegroup_config)?;
+            let rg_service = build_rolegroup_services(&druid, &rolegroup, &druid_tls_settings)?;
             let rg_configmap = build_rolegroup_config_map(
-                &rolegroup,
                 &druid,
+                &rolegroup,
                 rolegroup_config,
                 &zk_connstr,
                 opa_connstr.as_deref(),
                 s3_conn.as_ref(),
                 deep_storage_bucket_name.as_deref(),
                 &resources,
+                &druid_tls_settings,
             )?;
             let rg_statefulset = build_rolegroup_statefulset(
-                &rolegroup,
                 &druid,
+                &rolegroup,
                 rolegroup_config,
                 s3_conn.as_ref(),
                 &resources,
+                &druid_tls_settings,
             )?;
             cluster_resources
                 .add(client, &rg_service)
@@ -342,7 +375,12 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
 
 /// The server-role service is the primary endpoint that should be used by clients that do not perform internal load balancing,
 /// including targets outside of the cluster.
-pub fn build_role_service(role_name: &str, druid: &DruidCluster) -> Result<Service> {
+pub fn build_role_service(
+    druid: &DruidCluster,
+    role: &DruidRole,
+    tls_settings: &DruidTlsSettings,
+) -> Result<Service> {
+    let role_name = role.to_string();
     let role_svc_name = format!(
         "{}-{}",
         druid.metadata.name.as_ref().unwrap_or(&"druid".to_string()),
@@ -357,24 +395,15 @@ pub fn build_role_service(role_name: &str, druid: &DruidCluster) -> Result<Servi
             .with_recommended_labels(
                 druid,
                 APP_NAME,
-                druid_version(druid)?,
+                druid.version(),
                 CONTROLLER_NAME,
-                role_name,
+                &role_name,
                 "global",
             )
             .build(),
         spec: Some(ServiceSpec {
-            ports: Some(vec![ServicePort {
-                name: Some(CONTAINER_HTTP_PORT.to_string()),
-                port: DruidRole::from_str(role_name)
-                    .unwrap()
-                    .get_http_port()
-                    .into(),
-                target_port: Some(IntOrString::String(CONTAINER_HTTP_PORT.to_string())),
-                protocol: Some("TCP".to_string()),
-                ..ServicePort::default()
-            }]),
-            selector: Some(role_selector_labels(druid, APP_NAME, role_name)),
+            ports: Some(tls_settings.service_ports(role)),
+            selector: Some(role_selector_labels(druid, APP_NAME, &role_name)),
             type_: Some("NodePort".to_string()),
             ..ServiceSpec::default()
         }),
@@ -385,14 +414,15 @@ pub fn build_role_service(role_name: &str, druid: &DruidCluster) -> Result<Servi
 #[allow(clippy::too_many_arguments)]
 /// The rolegroup [`ConfigMap`] configures the rolegroup based on the configuration given by the administrator
 fn build_rolegroup_config_map(
-    rolegroup: &RoleGroupRef<DruidCluster>,
     druid: &DruidCluster,
+    rolegroup: &RoleGroupRef<DruidCluster>,
     rolegroup_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
     zk_connstr: &str,
     opa_connstr: Option<&str>,
     s3_conn: Option<&S3ConnectionSpec>,
     deep_storage_bucket_name: Option<&str>,
     resources: &RoleResource,
+    tls_settings: &DruidTlsSettings,
 ) -> Result<ConfigMap> {
     let role = DruidRole::from_str(&rolegroup.role).unwrap();
     let mut cm_conf_data = BTreeMap::new(); // filename -> filecontent
@@ -437,6 +467,10 @@ fn build_rolegroup_config_map(
                     DS_BUCKET.to_string(),
                     deep_storage_bucket_name.map(str::to_string),
                 );
+
+                // add tls encryption / auth properties
+                tls_settings.add_tls_config_properties(&mut transformed_config, &role);
+
                 let runtime_properties =
                     stackable_operator::product_config::writer::to_java_properties_string(
                         transformed_config.iter(),
@@ -479,7 +513,7 @@ fn build_rolegroup_config_map(
             .with_recommended_labels(
                 druid,
                 APP_NAME,
-                druid_version(druid)?,
+                druid.version(),
                 CONTROLLER_NAME,
                 &rolegroup.role,
                 &rolegroup.role_group,
@@ -500,9 +534,9 @@ fn build_rolegroup_config_map(
 ///
 /// This is mostly useful for internal communication between peers, or for clients that perform client-side load balancing.
 fn build_rolegroup_services(
-    rolegroup: &RoleGroupRef<DruidCluster>,
     druid: &DruidCluster,
-    _rolegroup_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
+    rolegroup: &RoleGroupRef<DruidCluster>,
+    tls_settings: &DruidTlsSettings,
 ) -> Result<Service> {
     let role = DruidRole::from_str(&rolegroup.role).unwrap();
 
@@ -515,7 +549,7 @@ fn build_rolegroup_services(
             .with_recommended_labels(
                 druid,
                 APP_NAME,
-                druid_version(druid)?,
+                druid.version(),
                 CONTROLLER_NAME,
                 &rolegroup.role,
                 &rolegroup.role_group,
@@ -524,20 +558,7 @@ fn build_rolegroup_services(
             .build(),
         spec: Some(ServiceSpec {
             cluster_ip: Some("None".to_string()),
-            ports: Some(vec![
-                ServicePort {
-                    name: Some(CONTAINER_HTTP_PORT.to_string()),
-                    port: role.get_http_port().into(),
-                    protocol: Some("TCP".to_string()),
-                    ..ServicePort::default()
-                },
-                ServicePort {
-                    name: Some(CONTAINER_METRICS_PORT.to_string()),
-                    port: DRUID_METRICS_PORT.into(),
-                    protocol: Some("TCP".to_string()),
-                    ..ServicePort::default()
-                },
-            ]),
+            ports: Some(tls_settings.service_ports(&role)),
             selector: Some(role_group_selector_labels(
                 druid,
                 APP_NAME,
@@ -555,68 +576,48 @@ fn build_rolegroup_services(
 ///
 /// The [`Pod`](`stackable_operator::k8s_openapi::api::core::v1::Pod`)s are accessible through the corresponding [`Service`] (from [`build_rolegroup_services`]).
 fn build_rolegroup_statefulset(
-    rolegroup_ref: &RoleGroupRef<DruidCluster>,
     druid: &DruidCluster,
+    rolegroup_ref: &RoleGroupRef<DruidCluster>,
     rolegroup_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
     s3_conn: Option<&S3ConnectionSpec>,
     resources: &RoleResource,
+    tls_settings: &DruidTlsSettings,
 ) -> Result<StatefulSet> {
-    // setup
     let role = DruidRole::from_str(&rolegroup_ref.role).context(UnidentifiedDruidRoleSnafu {
         role: rolegroup_ref.role.to_string(),
     })?;
-    let druid_version = druid_version(druid)?;
+    let druid_version = druid.version();
 
     // init container builder
-    let mut cb = ContainerBuilder::new(APP_NAME)
+    let mut cb_prepare = ContainerBuilder::new("prepare")
+        .context(FailedContainerBuilderCreationSnafu { name: "prepare" })?;
+    // druid container builder
+    let mut cb_druid = ContainerBuilder::new(APP_NAME)
         .context(FailedContainerBuilderCreationSnafu { name: APP_NAME })?;
     // init pod builder
     let mut pb = PodBuilder::new();
-    pb.metadata_builder(|m| {
-        m.with_recommended_labels(
-            druid,
-            APP_NAME,
-            druid_version,
-            CONTROLLER_NAME,
-            &rolegroup_ref.role,
-            &rolegroup_ref.role_group,
-        )
-    });
 
-    // add image
-    cb.image(container_image(druid_version));
+    // volume and volume mounts
+    tls_settings
+        .add_tls_volume_and_volume_mounts(&mut cb_prepare, &mut cb_druid, &mut pb)
+        .context(InvalidTlsConfigSnafu)?;
+    add_s3_volume_and_volume_mounts(s3_conn, &mut cb_druid, &mut pb)?;
+    add_config_volume_and_volume_mounts(rolegroup_ref, &mut cb_druid, &mut pb);
+    add_hdfs_cm_volume_and_volume_mounts(
+        &druid.spec.cluster_config.deep_storage,
+        &mut cb_druid,
+        &mut pb,
+    );
 
-    if let Some(s3_conn) = s3_conn {
-        if let Some(credentials) = &s3_conn.credentials {
-            pb.add_volume(credentials.to_volume("s3-credentials"));
-            cb.add_volume_mount("s3-credentials", S3_SECRET_DIR_NAME);
-        }
+    let prepare_container_command = tls_settings.build_tls_key_stores_cmd();
 
-        if let Some(tls) = &s3_conn.tls {
-            match &tls.verification {
-                TlsVerification::None {} => return S3TlsNoVerificationNotSupportedSnafu.fail(),
-                TlsVerification::Server(server_verification) => {
-                    match &server_verification.ca_cert {
-                        CaCert::WebPki {} => {}
-                        CaCert::SecretClass(secret_class) => {
-                            let volume_name = format!("{secret_class}-tls-certificate");
-
-                            let volume = VolumeBuilder::new(&volume_name)
-                                .ephemeral(
-                                    SecretOperatorVolumeSourceBuilder::new(secret_class).build(),
-                                )
-                                .build();
-                            pb.add_volume(volume);
-                            cb.add_volume_mount(&volume_name, format!("{CERTS_DIR}{volume_name}"));
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // add command
-    cb.command(role.get_command(s3_conn));
+    cb_prepare
+        .image("docker.stackable.tech/stackable/tools:0.2.0-stackable0")
+        .command(vec!["/bin/bash".to_string(), "-c".to_string()])
+        .args(vec![prepare_container_command.join(" && ")])
+        .image_pull_policy("IfNotPresent")
+        .security_context(SecurityContextBuilder::run_as_root())
+        .build();
 
     // rest of env
     let rest_env = rolegroup_config
@@ -630,54 +631,32 @@ fn build_rolegroup_statefulset(
             ..EnvVar::default()
         })
         .collect::<Vec<_>>();
-    cb.add_env_vars(rest_env);
 
-    // add ports
-    cb.add_container_port(CONTAINER_HTTP_PORT, role.get_http_port().into());
-    cb.add_container_port(CONTAINER_METRICS_PORT, DRUID_METRICS_PORT.into());
+    cb_druid.image(container_image(druid_version));
+    cb_druid.command(role.get_command(s3_conn));
+    cb_druid.image_pull_policy("IfNotPresent");
+    cb_druid.add_env_vars(rest_env);
+    cb_druid.add_container_ports(tls_settings.container_ports(&role));
+    // 10s * 30 = 300s to come up
+    cb_druid.startup_probe(tls_settings.get_tcp_socket_probe(30, 10, 30, 3));
+    // 10s * 1 = 10s to get removed from service
+    cb_druid.readiness_probe(tls_settings.get_tcp_socket_probe(10, 10, 1, 3));
+    // 10s * 3 = 30s to be restarted
+    cb_druid.liveness_probe(tls_settings.get_tcp_socket_probe(10, 10, 3, 3));
+    cb_druid.resources(resources.as_resource_requirements());
 
-    // config mount
-    cb.add_volume_mount("config", DRUID_CONFIG_DIRECTORY);
-    pb.add_volume(
-        VolumeBuilder::new("config")
-            .with_config_map(rolegroup_ref.object_name())
-            .build(),
-    );
-
-    // hdfs deep storage mount
-    if let DeepStorageSpec::HDFS(hdfs) = &druid.spec.deep_storage {
-        cb.add_volume_mount("hdfs", HDFS_CONFIG_DIRECTORY);
-        pb.add_volume(
-            VolumeBuilder::new("hdfs")
-                .with_config_map(&hdfs.config_map_name)
-                .build(),
-        );
-    }
-
-    // read write config folder
-    cb.add_volume_mount("rwconfig", RW_CONFIG_DIRECTORY);
-    pb.add_volume(
-        VolumeBuilder::new("rwconfig")
-            .with_empty_dir(Some(""), None)
-            .build(),
-    );
-
-    // readiness probe
-    let probe = Probe {
-        tcp_socket: Some(TCPSocketAction {
-            port: IntOrString::Int(role.get_http_port().into()),
-            ..Default::default()
-        }),
-        initial_delay_seconds: Some(30),
-        period_seconds: Some(5),
-        ..Default::default()
-    };
-    cb.readiness_probe(probe);
-    cb.resources(resources.as_resource_requirements());
-
-    let mut container = cb.build();
-    container.image_pull_policy = Some("IfNotPresent".to_string());
-    pb.add_container(container);
+    pb.add_init_container(cb_prepare.build());
+    pb.add_container(cb_druid.build());
+    pb.metadata_builder(|m| {
+        m.with_recommended_labels(
+            druid,
+            APP_NAME,
+            druid_version,
+            CONTROLLER_NAME,
+            &rolegroup_ref.role,
+            &rolegroup_ref.role_group,
+        )
+    });
     pb.security_context(PodSecurityContextBuilder::new().fs_group(1000).build()); // Needed for secret-operator
 
     Ok(StatefulSet {
@@ -719,12 +698,83 @@ fn build_rolegroup_statefulset(
     })
 }
 
-fn container_image(version: &str) -> String {
-    format!("docker.stackable.tech/stackable/druid:{}", version)
+fn add_hdfs_cm_volume_and_volume_mounts(
+    deep_storage_spec: &DeepStorageSpec,
+    cb_druid: &mut ContainerBuilder,
+    pb: &mut PodBuilder,
+) {
+    // hdfs deep storage mount
+    if let DeepStorageSpec::HDFS(hdfs) = deep_storage_spec {
+        cb_druid.add_volume_mount("hdfs", HDFS_CONFIG_DIRECTORY);
+        pb.add_volume(
+            VolumeBuilder::new("hdfs")
+                .with_config_map(&hdfs.config_map_name)
+                .build(),
+        );
+    }
 }
 
-pub fn druid_version(druid: &DruidCluster) -> Result<&str> {
-    Ok(&druid.spec.version)
+fn add_config_volume_and_volume_mounts(
+    rolegroup_ref: &RoleGroupRef<DruidCluster>,
+    cb_druid: &mut ContainerBuilder,
+    pb: &mut PodBuilder,
+) {
+    cb_druid.add_volume_mount("config", DRUID_CONFIG_DIRECTORY);
+    pb.add_volume(
+        VolumeBuilder::new("config")
+            .with_config_map(rolegroup_ref.object_name())
+            .build(),
+    );
+    cb_druid.add_volume_mount("rwconfig", RW_CONFIG_DIRECTORY);
+    pb.add_volume(
+        VolumeBuilder::new("rwconfig")
+            .with_empty_dir(Some(""), None)
+            .build(),
+    );
+}
+
+fn add_s3_volume_and_volume_mounts(
+    s3_conn: Option<&S3ConnectionSpec>,
+    cb_druid: &mut ContainerBuilder,
+    pb: &mut PodBuilder,
+) -> Result<()> {
+    if let Some(s3_conn) = s3_conn {
+        if let Some(credentials) = &s3_conn.credentials {
+            pb.add_volume(credentials.to_volume("s3-credentials"));
+            cb_druid.add_volume_mount("s3-credentials", S3_SECRET_DIR_NAME);
+        }
+
+        if let Some(tls) = &s3_conn.tls {
+            match &tls.verification {
+                TlsVerification::None {} => return S3TlsNoVerificationNotSupportedSnafu.fail(),
+                TlsVerification::Server(server_verification) => {
+                    match &server_verification.ca_cert {
+                        CaCert::WebPki {} => {}
+                        CaCert::SecretClass(secret_class) => {
+                            let volume_name = format!("{secret_class}-tls-certificate");
+
+                            let volume = VolumeBuilder::new(&volume_name)
+                                .ephemeral(
+                                    SecretOperatorVolumeSourceBuilder::new(secret_class).build(),
+                                )
+                                .build();
+                            pb.add_volume(volume);
+                            cb_druid.add_volume_mount(
+                                &volume_name,
+                                format!("{CERTS_DIR}/{volume_name}"),
+                            );
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    Ok(())
+}
+
+fn container_image(version: &str) -> String {
+    format!("docker.stackable.tech/stackable/druid:{}", version)
 }
 
 pub fn error_policy(_error: &Error, _ctx: Arc<Ctx>) -> Action {